refactor: room 패키지의 vo 패키지 분리와 dto 패키지 내 request/response 분리
