+++ backend/src/main/java/com/codesquad/airbnb/room/RoomService.java
@@ -6,9 +6,9 @@ import com.codesquad.airbnb.common.embeddable.GuestGroup;
 import com.codesquad.airbnb.common.embeddable.StayDate;
 import com.codesquad.airbnb.exception.ErrorCode;
 import com.codesquad.airbnb.exception.unchecked.NotFoundException;
-import com.codesquad.airbnb.room.dto.RoomDetailResponse;
-import com.codesquad.airbnb.room.dto.RoomSearCondition;
-import com.codesquad.airbnb.room.dto.RoomSearchResponse;
+import com.codesquad.airbnb.room.dto.request.RoomSearCondition;
+import com.codesquad.airbnb.room.dto.response.RoomDetailResponse;
+import com.codesquad.airbnb.room.dto.response.RoomSearchResponse;
 import com.codesquad.airbnb.room.entity.Room;
 import com.codesquad.airbnb.room.repository.RoomRepository;
 import java.util.List;
