+++ backend/src/main/java/com/codesquad/airbnb/room/dto/request/RoomSearCondition.java
@@ -1,4 +1,4 @@
-package com.codesquad.airbnb.room.dto;
+package com.codesquad.airbnb.room.dto.request;
 
 import com.codesquad.airbnb.common.embeddable.GuestGroup;
 import com.codesquad.airbnb.common.embeddable.Location;
