+++ backend/src/main/java/com/codesquad/airbnb/room/dto/response/RoomDetailResponse.java
@@ -1,27 +1,28 @@
-package com.codesquad.airbnb.room.dto;
+package com.codesquad.airbnb.room.dto.response;
 
 import com.codesquad.airbnb.common.embeddable.GuestGroup;
-import com.codesquad.airbnb.common.embeddable.Location;
 import com.codesquad.airbnb.common.embeddable.ReviewStat;
 import com.codesquad.airbnb.room.entity.Room;
 import com.codesquad.airbnb.room.entity.Room.RoomType;
 import com.codesquad.airbnb.room.entity.RoomDetail;
+import com.codesquad.airbnb.room.entity.RoomImage;
 import com.codesquad.airbnb.room.entity.embeddable.RoomGroup;
 import com.codesquad.airbnb.room.entity.embeddable.RoomOption;
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class RoomSearchResponse {
+public class RoomDetailResponse {
 
     private String name;
     private String address;
-    private String imagePath;
+    private String description;
     private RoomType type;
     private Integer price;
-    private Double longitude;
-    private Double latitude;
+    private List<String> imagePaths;
     private Double reviewScore;
     private Integer reviewCount;
     private Integer numberAdult;
@@ -34,22 +35,23 @@ public class RoomSearchResponse {
     private Boolean wirelessInternet;
     private Boolean airConditioner;
 
-    public static RoomSearchResponse from(Room room) {
+    public static RoomDetailResponse from(Room room) {
         RoomDetail detail = room.getDetail();
-        Location location = room.getLocation();
         ReviewStat review = room.getReview();
         GuestGroup guestGroup = detail.getGuestGroup();
         RoomGroup roomGroup = detail.getRoomGroup();
         RoomOption roomOption = detail.getOption();
+        List<String> imagePaths = room.getImages().stream()
+            .map(RoomImage::getPath)
+            .collect(Collectors.toList());
 
-        return new RoomSearchResponse(
+        return new RoomDetailResponse(
             room.getName(),
             room.getDistrict().getAddress(),
-            room.getImagePath(),
+            room.getDescription(),
             room.getType(),
             room.getPrice().getLodging(),
-            location.getLongitude(),
-            location.getLatitude(),
+            imagePaths,
             review.getScore(),
             review.getCount(),
             guestGroup.getNumberAdult(),
