+++ backend/src/test/java/com/codesquad/airbnb/room/RoomServiceTest.java
@@ -15,10 +15,10 @@ import com.codesquad.airbnb.member.Member;
 import com.codesquad.airbnb.member.Member.MemberRole;
 import com.codesquad.airbnb.reservation.Reservation;
 import com.codesquad.airbnb.reservation.Reservation.ReservationState;
-import com.codesquad.airbnb.room.dto.RoomSearCondition;
-import com.codesquad.airbnb.room.dto.RoomSearCondition.PriceRange;
-import com.codesquad.airbnb.room.dto.RoomSearCondition.Radius;
-import com.codesquad.airbnb.room.dto.RoomSearchResponse;
+import com.codesquad.airbnb.room.dto.request.RoomSearCondition;
+import com.codesquad.airbnb.room.dto.request.RoomSearCondition.PriceRange;
+import com.codesquad.airbnb.room.dto.request.RoomSearCondition.Radius;
+import com.codesquad.airbnb.room.dto.response.RoomSearchResponse;
 import com.codesquad.airbnb.room.entity.Room;
 import com.codesquad.airbnb.room.entity.Room.RoomType;
 import com.codesquad.airbnb.room.entity.RoomDetail;
